+++ code/graphx/src/test/scala/com/bd/propogation/ic/ConnectedComponentsTest.scala
@@ -16,7 +16,7 @@ class ConnectedComponentsTest extends SparkTestBase {
     val size = 4
     val weight = 1
     val vertices = createVertices(size)
-    var edges: List[Edge[Int]] = List()
+    var edges: List[Edge[Double]] = List()
     edges ::= Edge(1, 2, weight)
     edges ::= Edge(1, 3, weight)
     edges ::= Edge(1, 4, weight)
